DD-1018 part way through refactor to use a separate endpoint and table for safeguarding data.
